I init useInitialMessages hook.
